# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface GeminiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface GeminiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface GeminiConversationInput {
  title: string;
}

export interface GeminiMessageInput {
  content: string;
  arenaRole?: string;
  arenaFeature?: string;
  selectedLanguage?: string;
}

export interface GeminiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: GeminiMessage[];
}

export interface GeminiError {
  error: string;
}

export type OrganizerSummaryInputMetrics = { [key: string]: unknown };

export interface OrganizerSummaryInput {
  metrics?: OrganizerSummaryInputMetrics;
}

export interface OrganizerSummaryOutput {
  summary: string;
  recommendations: string[];
  raw?: string;
}

export interface AnnouncementInput {
  situation: string;
  language?: string;
}

export interface AnnouncementOutput {
  announcement: string;
  language: string;
}