# lib/api-zod/src/generated/types/geminiConversationWithMessages.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GeminiMessage } from './geminiMessage';

export interface GeminiConversationWithMessages {
  id: number;
  title: string;
  createdAt: Date;
  messages: GeminiMessage[];
}