# lib/api-zod/src/generated/types/geminiMessageInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GeminiMessageInput {
  content: string;
  arenaRole?: string;
  arenaFeature?: string;
  selectedLanguage?: string;
}