# lib/api-zod/src/generated/types/organizerSummaryOutput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrganizerSummaryOutput {
  summary: string;
  recommendations: string[];
  raw?: string;
}